requirements.txt
flask
flask-cors
google-cloud-speech
gunicorn